TODO: update de cÂmera

+++ include/State.h
@@ -24,7 +24,7 @@ class State {
     void LoadAssets(); // pré-carregar os assets do state do jogo
     // carregar imagens/fontes/músicas às suas variáveis aqui sempre
     // que for possível.
-    void Update(double dt);
+    void Update(float dt);
     void Render();
 };
